requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
astroid==2.4.2
brotlipy 
certifi==2020.6.20
cffi 
charset-normalizer 
click==8.1.2
colorama==0.4.4
cryptography 
cycler==0.11.0
distlib==0.3.1
Flask==2.1.1
fonttools==4.31.2
gunicorn==20.1.0
idna 
imgaug==0.4.0
importlib-metadata==4.11.3
isort==5.7.0
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
lazy-object-proxy==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.1
mccabe==0.6.1
numpy 
olefile==0.46
opencv-python==4.5.3.56
packaging==21.3
pandas==1.4.2
Pillow 
pycparser 
pylint==2.6.0
pyOpenSSL 
pyparsing==3.0.7
PySocks 
python-dateutil==2.8.2
pytz==2022.1
requests 
six==1.15.0
timm==0.5.4
toml==0.10.2
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
typing_extensions 
urllib3 
virtualenv-clone==0.5.4
Werkzeug==2.1.1
win-inet-pton 
wrapt==1.12.1
zipp==3.8.0
